Temporary broken state: Removed Hole/Fragment results functionality.
Can't type code now, hence the WYTIWYS tests fail.
Towards adding completely new UX design.

Starting the change by removing this code,
as this is some of the worst parts of our code base.
It will be better to re-implement the new UX design with a fresh start.

Also, the hole's design was a search-menu design, which isn't relevant to the new UX design,
so replaced it with a simple "_", which should be familiar to Haskellers.
We'll need to decide what holes look like!